requirements.txt
fastapi
uvicorn[standard]
anthropic
google-generativeai
huggingface_hub
python-multipart
python-dotenv
pymupdf
python-docx
pillow
requests
